renamed tag_script to lexicomb and added License
A lexicon is a corpus of terms and definitions. by defining one or
more tags, we define a lexicon for our language. That is part of the
reason for the lexi in lexicomb. Of course, lexicomb also uses a
lexical analayzer and parser combinator, so that is the other reason
for the name.

Added a copy of the official Apache 2.0 license

changed namespace from bbp to bbpyp since bbp is already being used on
pypi. bbpyp stands for Blogger Bust Python Projects

Signed-off-by: BloggerBust <trevor.wilson@bloggerbust.ca>